backend/requirements.txt
firebase-admin==6.6.0
playwright==1.53.0
python-dotenv==1.1.1

